+++ include/loss.h
@@ -5,14 +5,14 @@
 
 class Loss {
  protected:
-  float loss;  // value of loss
+  double loss;  // value of loss
   Matrix grad_bottom;  // gradient w.r.t input
 
  public:
   virtual ~Loss() {}
 
   virtual void evaluate(const Matrix& pred, const Matrix& target) = 0;
-  virtual float output() { return loss; }
+  virtual double output() { return loss; }
   virtual const Matrix& back_gradient() { return grad_bottom; }
 };
