+++ include/sgd.h
@@ -6,12 +6,12 @@
 
 class SGD : public Optimizer {
  private:
-  float momentum;  // momentum factor (default: 0)
+  double momentum;  // momentum factor (default: 0)
   bool nesterov;  // enables Nesterov momentum (default: False)
-  std::unordered_map<const float*, Vector> v_map;  // velocity
+  std::unordered_map<const double*, Vector> v_map;  // velocity
 
  public:
-  explicit SGD(float lr = 0.01, float decay = 0.0, float momentum = 0.0,
+  explicit SGD(double lr = 0.01, double decay = 0.0, double momentum = 0.0,
                bool nesterov = false) : Optimizer(lr, decay),
                momentum(momentum), nesterov(nesterov) {}
